修改部分bug

+++ engine_center/src/main/java/com/bilibili/engine_center/startup/AutoSpeed.java
@@ -41,7 +41,7 @@ public class AutoSpeed {
     }
 
 
-    public void init(Context context, long coldStartTime){
+    public void init(Context context, int coldStartTime){
         this.application = context.getApplicationContext();
         if (this.coldStartTime < 0){
             this.coldStartTime = coldStartTime;
